android_looper_sys::ALooper_callbackFunc

type ALooper_callbackFunc = Option<unsafe extern fn(fd: c_int, events: c_int, data: *mut c_void) -> c_int>;

For callback-based event loops, this is the prototype of the function that is called when a file descriptor event occurs. It is given the file descriptor it is associated with, a bitmask of the poll events that were triggered (typically ALOOPER_EVENT_INPUT), and the data pointer that was originally supplied.

Implementations should return 1 to continue receiving callbacks, or 0 to have this file descriptor and callback unregistered from the looper.